riotboot: fix concurrent builds
Building an app `foo` for board, say, `same54-xpro` that depends on
riotbuilds resulted in two `$(BINDIR)`s to be created:

1. `path/to/app/foo/bin/same54-xpro`
2. `$(RIOT_BASE)/bootloaders/riotboot/bin/same54-xpro`

In the folder, `riotboot` would not be build for `same54-xpro`.

If now a second app `bar` is build for the same board, this will
build in:

1. `path/to/app/bar/bin/same54-xpro`
2. `$(RIOT_BASE)/bootloaders/riotboot/bin/same54-xpro`

Now the two concurrent builds will overwrite the `riotboot` compilation
of each other.

This commit fixes the issue by building `riotboot` instead in:

1. `path/to/app/foo/bin/same54-xpro/riotboot_files`
2. `path/to/app/bar/bin/same54-xpro/riotboot_files`

This way the two concurrent builds can work independently of each other
and wont overwrite each other's build files.