build-system: generate buildinfo/{app,board,cpu,riotver}.h
This changse the build system to expose the board name, the app name,
the CPU name, and the RIOT version via dynamically generated headers
rather than via `CFLAGS += -D`.

The intended goal is to to reduce the entropy of the input of ccache
with the hope that for platform independent code, the compiler input
to ccache will be often identical for related boards and similar apps.

The downside is that code that previously dependent on the macros being
predefined by the environment has not to explicitly include the suitable
dynamically created header(s).

One new macro has also been added: `-DRIOT_OS=1`. Previously a lot of
code was checking for `RIOT_VERSION` to compile differently under e.g.
Linux and RIOT. The code has been updated to check for `RIOT_OS` instead.

Co-authored-by: crasbe <crasbe@gmail.com>