build-system: Pass `-D` flags with `@file` argument
This drops the generation of `riotbuild.h` and instead just passes the
`-D...` / `-U...` flags via a command line options file to GCC/clang
instead.

The old `riotbuild.h` now is moved to `core/include` and only contains
deprecated aliases and documentation. It is still injected with
`-include riotbuild.h`, as otherwise the aliases for dynamic macros
injected with `-D` would not be visible.